+++ js/catalogue.js
@@ -34,7 +34,7 @@ function pageBlock(action, block, text) {
 }
 
 function addToBasket(good_id, input, response_field) {
-	var quantity = parseInt(document.getElementById(input).value);
+	var quantity = parseFloat(document.getElementById(input).value);
 
 	if(quantity > 0) {
 		$.ajax({
