+++ js/item.js
@@ -71,7 +71,7 @@ function saveGoodPrice(good_id, block, currency, unit, rate) {
 }
 
 function addToBasket(good_id, input, response_field) {
-	var quantity = parseInt(document.getElementById(input).value);
+	var quantity = parseFloat(document.getElementById(input).value);
 
 	if(quantity > 0) {
 		$.ajax({
